+++ poc_mobile_verifier/src/lib.rs
@@ -3,13 +3,13 @@ mod error;
 mod heartbeats;
 mod ingest;
 mod mobile;
-mod reward_share;
+mod owner_shares;
 mod settings;
 mod speedtests;
 
 pub mod cli;
 pub mod scheduler;
-pub mod subnetwork_rewards;
+// pub mod subnetwork_rewards;
 pub mod verifier;
 
 pub use error::{Error, Result};
